requirements.txt
setuptools>=69
wheel
numpy==2.2.5
pillow==9.5.0
flask==2.3.3
werkzeug==2.3.7
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
torchvision==0.18.1+cpu
gunicorn
